requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
lxml
celery
redis
mysqlclient
pymysql
cryptography
python-decouple
pytz